Fix serverless loader for API routes
